Some other import modules and test changes
